+++ wsei-todo-app-master/src/screens/Home/index.tsx
@@ -1,14 +1,17 @@
 import React, { FC } from 'react';
-import {  View, ScrollView } from 'react-native';
+import {  View, ScrollView,Image} from 'react-native';
 import styled from 'styled-components/native';
 import { useNavigation } from '@react-navigation/native';
-
+import {Dimensions} from 'react-native';
 import Colors from '../../constans/Colors';
+import {  Text,  StyleSheet, } from "react-native";
 
 const WelcomeText = styled.Text`
     margin: 10px 20px;
     font-size: 16px;
     color: ${Colors.black};
+    text-align: center;
+   
 `;
 
 const CustomImage = styled.Image`
@@ -17,6 +20,12 @@ const CustomImage = styled.Image`
     height: 80px;
     margin: 50px;
 `;
+const styles = StyleSheet.create({
+    
+    
+    tloImage: {position:"absolute", height:Dimensions.get('window').height ,opacity:0.3,    }
+   
+  });
 
 interface IHomeProps { }
 
@@ -24,24 +33,30 @@ const Home: FC<IHomeProps> = (props) => {
     const navigation = useNavigation();
 
     return ( 
+        <View>
+        <View>
+        <Image
+            source={require('../../assets/video.gif')}
+            style={styles.tloImage}
+             />
+             
+         </View>
         <ScrollView>
+        
         <View>
         <CustomImage
-                source={require('../../assets/logo.png')}
+                source={require('../../assets/images.jpg')}
             />
-            <WelcomeText>Spotkanie z Wilkami będzie dla Lewandowskiego okazją do wyrównania także innego rekordu – zdobycia bramek przeciwko szesnastu klubom w trakcie jednego sezonu. Do tej pory dwukrotnie dokonał tego Gerd Mueller, raz taka sztuka udała się Ailtonowi. Polak ma na rozkładzie piętnaście drużyn, najlepiej radził sobie przeciwko Schalke zdobywając cztery bramki.
+            <WelcomeText>Powstanie filmu wiązało się z odkryciem niedoskonałości ludzkiego wzroku i jego opóźnień w przetwarzaniu obrazu. W związku z tym powstały urządzenia „oszukujące” wzrok, które ciąg szybko zmieniających się nieruchomych obrazków przeobrażały w ruchomy obraz. Najwcześniejsze z nich to camera obscura; w XVIII i XIX w. powstały również takie urządzenia, jak fenakistiskop (1832 r.), zoetrop (1834 r.) i praksinoskop (1877 r.). Owe praktyki dziś zalicza się do wczesnej historii kina i nazywa jej prehistorią. W 1877 r. Hannibal Goodwin opracował taśmę celuloidową, co znacząco przyspieszyło rozwój techniki filmowej. Pierwsze, prymitywne filmy powstały orientacyjnie w latach 1887 i 1888, a ich autorem był francuski wynalazca Louis Le Prince; do innych pionierów sztuki filmowej należeli Max Skladanowsky oraz William Friese-Greene.
 
-Dogonić legendy
-
-31-letni napastnik w tym sezonie jest w niesamowitej formie. Jeśli zdoła utrzymać ją w następnych rozgrywkach, uda mu się dogonić legendy. Obecnie Robert Lewandowski jest trzecim najskuteczniejszym graczem Bundesligi w historii z 235 bramkami na koncie. Żaden z grających obecnie piłkarzy nie może równać się z napastnikiem Bayernu – lepsi od niego są tylko Klaus Fischer i Gerd Mueller. 365 bramek innej legendy Die Roten wydaje się poza czyimkolwiek zasięgiem, ale już do Fischera Polakowi brakuje ”zaledwie” 33 trafień – tyle Lewandowskiemu udało się zdobyć w obecnym sezonie.
-
-Fascynująca może być też walka o miejsce na podium najlepszych strzelców Ligi Mistrzów. Obecnie trzecie miejsce wśród najskuteczniejszych graczy tych rozgrywek zajmuje Raul. Osiem bramek mniej mają Robert Lewandowski i Karim Benzema. W znacznie lepszej sytuacji jest Polak, którego Bayern jest jedną nogą w ćwierćfinale Ligi Mistrzów po wygraniu z Chelsea 3:0. Francuz w tym sezonie może mieć mniej okazji na poprawienie imponujących statystyk, bowiem Real Madryt w pierwszym spotkaniu 1/8 finału przegrał na własnym stadionie z Manchesterem City 1:2.
+Pod koniec XIX w. powstało wiele eksperymentalnych urządzeń do wyświetlania filmów (m.in. kinetoskop autorstwa Thomasa Alvy Edisona z 1891 r.). Na początku XX w., w latach 1900-1915, w Stanach Zjednoczonych i Europie pojawiło się łącznie kilkadziesiąt rodzajów urządzeń do rejestracji i demonstracji filmów ze zsynchronizowanym dźwiękiem. Przykładami mogą być: chronograf (L. Gaumont, 1902 r.), biofon (O. Messter, 1903 r.) i kamerofon (USA, 1908 r.). Przy pomocy tych urządzeń, w Europie (w Niemczech, Francji, Włoszech i Danii) do 1914 r., powstało około 1500 krótkich filmów udźwiękowionych. Wszystkie te próby nie przyniosły jednak ostatecznie rezultatów zadowalających z technicznego punktu widzenia. W latach 1907-1913 kilkadziesiąt takich urządzeń skonstruowano i wprowadzono na rynek amerykański, z takim skutkiem jak w Europie.
 
 </WelcomeText>
             
             
         </View>
         </ScrollView>
+        </View>
     );
 };